Title: 앞에서부터 삭제하기 2, Time: 1088ms, Memory: 2MB, Status: Time Limit Exceed - Codetree

+++ 260305/앞에서부터 삭제하기 2/delete-it-from-the-beginning-2.cpp
@@ -34,7 +34,7 @@ int main() {
             total += (-pq.top());
             pq.pop();
         }
-        ans = max(ans, total / pqs );
+        ans = max(ans, (double)total / pqs );
     }
     cout << fixed << setprecision(2) << ans ;
